// client/src/app/models/program-outcomes.ts
export class ProgramOutcomes {
  _id: string | undefined;
  curriculumId: string | undefined;
  curriculumName: string | undefined;
  poCode: string | undefined; // batch outcome code
  // poCode: string | undefined;
  poCodeStatement: string | undefined;
  // poCodeStatement: string | undefined;
  createdAt: Date | undefined;
  updateAt: Date | undefined;
  cognitiveDomain: string[] | undefined;
}